requirements.txt
beautifulsoup4==4.11.1
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
Flask==2.2.2
future==0.18.2
googlemaps==4.6.0
idna==3.4
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
line-bot-sdk==1.18.0
MarkupSafe==2.1.1
requests==2.28.1
soupsieve==2.3.2.post1
typing-extensions==4.4.0
urllib3==1.26.12
Werkzeug==2.2.2
zipp==3.10.0
